// src/services/hooks/sales/mutations/useSetDraft/interface.ts
export interface Draft {
  serviceType: string;
  subscribe: boolean;
  customer: string;
  payTerm: string;
  saleDate: string;
  invoiceScheme: string;
  invoiceNo?: string;
  document?: File | null;
  salesOrder?: string;
  selectTable: string;
  serviceStaff: string;
  productName: string;
  sku: string;
  quantity: string;
  unitPrice: string;
  discount: string;
  priceIncTax: string;
  subtotal: string;
  discountType: string;
  discountAmount: string;
  orderTax: string;
  sellNote?: string;
  shippingDetails: string;
  shippingAddress: string;
  shippingCharges: string;
  shippingStatus: string;
  deliveredTo: string;
  deliveryPerson?: string;
  shippingDocument?: File | null;
  expenseName: string;
  amount: string;
}